requirements.txt
fastapi
uvicorn
yfinance
langchain==0.1.20
langchain-openai==0.1.*
langchain-community
langgraph
pandas
requests
python-dotenv 
feedparser
apscheduler
pandas_ta
matplotlib